feat: LPDO iteration utilities
Define the `iterate` method for the `LPDO` type so that native Julia
features such as `for t in v ... end` where `v` is an `LPDO`, as well as
many functions in `Base` that require an iterator, can be used
straightforwardly on LPDOs. This requires adding some other methods that
work in the background, such as `keys`, `values`, and so on, that in
this case usually are forwarded to the underlying `data` matrix.